Extract ApolloServerPluginInlineTrace from EngineReportingAgent
The plugin itself doesn't get changed at all. The default logic for enabling it
changes slightly. Previously, if you configure an API key, then there was no
way (explicit or implicit) to enable inline federated tracing, and if you had a
federated schema it would log a warning.

Now, you can enable or disable it explicitly with your own
ApolloServerPluginInlineTrace or ApolloServerPluginInlineTraceDisabled. (It
doesn't have to be a federated service; we can imagine other consumers of the
inline trace data in the future.) If neither of these plugins are installed and
you are a federated service, then inline tracing is enabled (whether or not
there is an API key configured), and an info message is logged.  This means you
could have an API key configured for some other reason (eg operation registry?)
and still use federated tracing, which was previously impossible.